Moving file creator to separate class
