+++ Api.Services/Interfaces/IFileManager.cs
@@ -1,0 +1,10 @@
+﻿
+namespace Api.Services.Interfaces
+{
+    public interface IFileManager
+    {
+        bool CreateFile();
+
+        bool DelteFiles();
+    }
+}
