+++ Api.Services/MQBrokerFile.cs
@@ -24,37 +24,5 @@ namespace Api.Services
         {
             throw new NotImplementedException();
         }
-
-        public void CreateFiles()
-        {
-            var input = _options.Value.FilesRoute + _options.Value.InputData;
-            var output = _options.Value.FilesRoute + _options.Value.OutputData;
-            try
-            {
-                File.Create(input).Dispose();
-                File.Create(output).Dispose();
-            }
-            catch (Exception ex)
-            {
-                throw ex;
-            }
-
-        }
-
-        public void DeleteFiles()
-        {
-            var files = Directory.GetFiles(_options.Value.FilesRoute);
-            try
-            {
-                foreach(var file in files)
-                {
-                    File.Delete(file);
-                }
-            }
-            catch(Exception ex)
-            {
-                throw ex;
-            }
-        }
     }
 }
